Adapt test

+++ chatserver/src/main/java/com/wyssengineeringai/chatserver/entity/Message.java
@@ -1,7 +1,11 @@
 package com.wyssengineeringai.chatserver.entity;
 
+import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
+import org.springframework.data.relational.core.mapping.Column;
+
+import java.time.Instant;
 
 @Data 
 @Table(name = "messages")
@@ -12,10 +16,10 @@ public class Message {
     private String content;
 
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column("chat_room_id")
-    private Integer chatRoomId;
+    private Long chatRoomId;
 
     @Column("created_at")
     private Instant createdAt;
